backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
google-auth==2.38.0
google-auth-oauthlib==1.2.1
idna==3.10
itsdangerous==2.2.0
kombu==5.5.1
oauthlib==3.2.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
